+++ jest/__mocks__/*/cartridge/scripts/checkout/checkoutHelpers.js
@@ -5,3 +5,13 @@ export const createOrder = jest.fn(() => ({
   orderToken: 'mocked_orderToken',
 }));
 export const calculatePaymentTransaction = jest.fn(() => ({ error: false }));
+export const validateCreditCard = jest.fn(() => ({ creditCardErrors: 'mockedCreditCardErrors' }));
+export const savePaymentInstrumentToWallet = jest.fn(() => ({
+  creditCardHolder: "mockedCardHolder",
+  maskedCreditCardNumber: "mockedCardNumber",
+  creditCardType: "mockedCardType",
+  creditCardExpirationMonth: "mockedExpirationMonth",
+  creditCardExpirationYear: "mockedExpirationYear",
+  UUID: "mockedUUID",
+  creditCardNumber: "mockedCardNumber"
+}));
