+++ jest/__mocks__/*/cartridge/scripts/util/array.js
@@ -1,0 +1,1 @@
+export const find = jest.fn((str, callback) => str);
