+++ src/cartridges/int_adyen_SFRA/cartridge/scripts/hooks/payment/processor/middlewares/index.js
@@ -1,2 +1,4 @@
 module.exports.authorize = require('./authorize');
 module.exports.handle = require('./handle');
+module.exports.processForm = require('./processForm');
+module.exports.savePaymentInformation = require('./savePaymentInformation');
